requirements.txt
streamlit
langchain
langchain-community
langchain-core
langchain-chroma
langchain-groq
langchain-openai
chromadb
sentence-transformers
pandas
python-dotenv
python-docx
PyMuPDF
Pillow
pytesseract
openpyxl
onnxruntime